code refinements made

+++ invertedai_cpp/invertedai/location_info_response.h
@@ -20,7 +20,7 @@ private:
   std::string osm_map_;
   Point2d map_origin_;
   Point2d rendering_center_;
-  int rendering_fov_;
+  float rendering_fov_; 
   std::vector<StaticMapActor> static_actors_;
   json body_json_;
 
@@ -65,8 +65,14 @@ public:
    * Lists traffic lights with their IDs and locations.
    */
   std::vector<StaticMapActor> static_actors() const;
+    /**
+   * Gets the map center
+   */
   Point2d rendering_center() const;
-  int rendering_fov() const;
+    /**
+   * Gets the map fov 
+   */
+  float rendering_fov() const;
 
   // setters
   /**
